Move recipes.json and update how it's referenced (#75)
## Background

Now that I have a way to host mumsmums, I want to change how the db handling is done - I no longer want to check in the db since I'll look to make changes to the deployed version. 

However, before we do that; I want to do some cleanup. The recipes.json should be used as the source/seed for the db if it doesn't already exist - let's move it to the server resources folder as I'll want to reference it from the init stage of the db setup later. 

## Change

- Move recipes.json to server resource folder
- Update how we make recipes.json available to the running app through Bazel

## Test plan

- It builds
- It tests
- Ran the db initialize script without issue