Don't check in db file (#79)
## Background

Now that I have a deployed version of mumsmums; I no longer want to check in the db file as it should be persisted by host. Therefore, restructure the code to create/persist the db file if it doesn't exist (using the recipes.json source) - or simply use it if it's already found on disk. 

## Change

- Remove sqlite/mumsmums.db and add it to .gitignore 
- Restructure mumsmums to initialize the db if it's empty 

## Test plan

- It builds and runs locally
- It builds and runs locally in Docker
- It builds and runs in Docker after deploying