Add export script (#90)
## Background

Since I'm developing and adding recipes locally (using a local admin page), that modifies my _local_ db, but obviously not the deployed one. 

Therefore, it would be handy to use a script to export the local DB to the source recipes.json file - this one can then be used on the deployed version to add new recipes.

## Change

- Add export script
- Add a pre-commit hook for json 
- Update recipes.json (formatting is now done as per the pre-commit hook, instead of having been formatted in sublime text)

## Test plan

- It builds
- It tests
- It formats