Add skeleton for search panel (#94)
## Background

I want to implement search and filtering - taking the first steps towards this by introducing a SearchPanel (hidden behind the `SEARCH` "feature flag"). 

This introduces a component that opens a panel when you click the search button, and shows an input field. The input field does not yet do anything, this is simply covering the UI/layout, making sure you can scroll with it expanded etc:

<img width="429" height="608" alt="Screenshot 2025-12-15 at 14 56 51" src="https://github.com/user-attachments/assets/0e54922a-4586-4c75-9864-f13e826b2d30" />

## Change

- Add SearchPanel
- Update Nav to ensure it's pinned to the top (important when scrolling when the search panel is open) 

## Test plan

- It builds
- It tests